requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.7.1
deprecation==2.1.0
dnspython==2.6.1
docopt==0.6.2
elementpath==4.4.0
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flaskapp==1.0.5
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grip==4.6.2
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jwcrypto==1.5.6
jwt==1.3.1
Markdown==3.5.2
MarkupSafe==2.1.5
mongo==0.2.0
oauthlib==3.2.2
packaging==24.0
path-and-address==2.0.1
pathlib==1.0.1
pillow==10.2.0
protobuf==4.25.3
py-healthcheck==1.10.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.2
pyOpenSSL==24.1.0
pyotp==2.9.0
pyparsing==3.1.2
pypng==0.20220715.0
pysaml2==7.5.0
python-dateutil==2.9.0.post0
python-keycloak==3.9.1
pytz==2024.1
qrcode==7.4.2
qrcode-terminal==0.8
requests==2.31.0
requests-oauthlib==1.4.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.1
xmlschema==2.5.1
